# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Это API (swagger 2.0) системы автоматизации психолого-педагогическ\
    ого сопровождения субъектов образовательной деятельности для школы №18."
  version: "1.0.7"
  title: "PsychoSys18"
  contact:
    email: "antony.nor@yandex.ru"
host: "api-ps18.herokuapp.com"
basePath: "/"
...
schemes:
- "https"
paths:
  /downloadFiles:
    post:
      tags:
      - "files"
      summary: "Скачивание файлов"
      operationId: "downloadFiles"
      consumes:
      - "application/json"
      produces:
      - "multipart/form-data"
      parameters:
      - in: "body"
        name: "body"
        description: "Массив ID файлов"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/body"
      responses:
        200:
          description: "Статус запроса, файлы и дополнительная информация"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Files"
    x-swagger-router-controller: "files"
  /getFiles:
    post:
      tags:
      - "files"
      summary: "Получение информации о файлах тестов/рез-ов тестирований/ИПК"
      operationId: "getFiles"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Либо ID студента, либо ID рез. тестирования, либо ID самого\
          \ тестирования"
        required: true
        schema:
          $ref: "#/definitions/body_1"
      responses:
        200:
          description: "Статус запроса и файлы с информацией о них"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_1_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Files"
    x-swagger-router-controller: "files"
  /addFiles:
    post:
      tags:
      - "files"
      summary: "Добавление файлов тестов/рез-ов тестирований/ИПК"
      operationId: "addFiles"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Либо ID студента, либо ID рез. тестирования, либо ID самого\
          \ тестирования и файлы"
        required: true
        schema:
          $ref: "#/definitions/body_2"
      responses:
        200:
          description: "Статус запроса, имена файлов, которые удалось прикрепить и\
            \ их ID"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_2_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Files"
    x-swagger-router-controller: "files"
  /updFiles:
    post:
      tags:
      - "files"
      summary: "Изменение файлов тестов и/или рез-ов тестирований и/или ИПК"
      operationId: "updFiles"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID файла, его имя и путь"
        required: true
        schema:
          $ref: "#/definitions/body_3"
      responses:
        200:
          description: "Статус запроса и обновленная информация о файле"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/inline_response_200_1_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Files"
    x-swagger-router-controller: "files"
  /delFiles:
    post:
      tags:
      - "files"
      summary: "Удаление файлов тестов и/или рез-ов тестирований и/или ИПК"
      operationId: "delFiles"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Массив ID файлов"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/body"
      responses:
        200:
          description: "Статус запроса и рузультат удаления по каждому файлу"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_4_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Files"
    x-swagger-router-controller: "files"
...
  /registerEmployees:
    post:
      tags:
      - "auth"
      summary: "Регистрация сотрудников из xlsx"
      operationId: "registerEmployees"
      consumes:
      - "application/octet-stream"
      produces:
      - "application/octet-stream"
      parameters:
      - in: "body"
        name: "body"
        description: "Файл в base64"
        required: true
        schema:
          type: "string"
          format: "base64"
      responses:
        200:
          description: "Статус запроса и файл с результатами регистрации."
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/inline_response_200_18_payload"
        401:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Auth"
    x-swagger-router-controller: "auth"
  /registerStudents:
    post:
      tags:
      - "auth"
      summary: "Регистрация учеников из xlsx"
      operationId: "registerStudents"
      consumes:
      - "application/octet-stream"
      produces:
      - "application/octet-stream"
      parameters:
      - in: "body"
        name: "body"
        description: "Файл в base64"
        required: true
        schema:
          type: "string"
          format: "base64"
      responses:
        200:
          description: "Статус запроса и файл с результатами регистрации."
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/inline_response_200_18_payload"
        401:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Auth"
    x-swagger-router-controller: "auth"
  /confirmParentReg:
    post:
      tags:
      - "auth"
      summary: "Подтверждение регистрации родителей"
      operationId: "confirmParentReg"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID родителя и статус подтверждения"
        required: true
        schema:
          $ref: "#/definitions/body_23"
      responses:
        200:
          description: "Статус выполнения запроса."
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Auth"
    x-swagger-router-controller: "auth"
  /getListConfirmReg:
    post:
      tags:
      - "auth"
      summary: "Получить список запросов на регистрацию"
      operationId: "getListConfirmReg"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Статус запроса и список запросов на регистрацию."
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_19_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Auth"
    x-swagger-router-controller: "auth"
  /autoCheckParentReg:
    post:
      tags:
      - "auth"
      summary: "Автопроверка подтверждения регистрации"
      operationId: "autoCheckParentReg"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID родителя"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/body_24"
      responses:
        200:
          description: "Статус запроса и результаты автопроверки подтверждения"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_20_payload"
        401:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Auth"
    x-swagger-router-controller: "auth"
...
  /getEmpPosts:
    post:
      tags:
      - "posts"
      summary: "Получить список должностей сотрудника"
      operationId: "getEmpPosts"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID сотрудника"
        required: true
        schema:
          $ref: "#/definitions/body_33"
      responses:
        200:
          description: "Список должностей сотрудника"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/Post"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Posts"
    x-swagger-router-controller: "posts"
  /addEmpPosts:
    post:
      tags:
      - "posts"
      summary: "Добавить должности сотруднику"
      operationId: "addEmpPosts"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID сотрудника и список ID должностей"
        required: true
        schema:
          $ref: "#/definitions/body_34"
      responses:
        200:
          description: "Статус выполнения запроса"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Posts"
    x-swagger-router-controller: "posts"
  /delEmpPosts:
    post:
      tags:
      - "posts"
      summary: "Удалить должности сотрудника"
      operationId: "delEmpPosts"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID сотрудника и список ID должностей"
        required: true
        schema:
          $ref: "#/definitions/body_35"
      responses:
        200:
          description: "Статус выполнения операции."
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Posts"
    x-swagger-router-controller: "posts"
  /getAllPosts:
    post:
      tags:
      - "posts"
      summary: "Получить список должностей"
      operationId: "getAllPosts"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Список должностей"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/Post"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Posts"
    x-swagger-router-controller: "posts"
  /addPost:
    post:
      tags:
      - "posts"
      summary: "Добавить должности"
      operationId: "addPost"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Список должностей"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/body_36"
      responses:
        200:
          description: "Статус выполнения запроса"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Posts"
    x-swagger-router-controller: "posts"
  /delPost:
    post:
      tags:
      - "posts"
      summary: "Удалить должности"
      operationId: "delPost"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Список ID должностей"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/signup_emp_data_pst_arr"
      responses:
        200:
          description: "Статус выполнения операции."
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Posts"
    x-swagger-router-controller: "posts"
securityDefinitions:
  Basic:
    type: "basic"
  JWT:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  People:
    type: "object"
    required:
    - "pepl_birthday"
    - "pepl_first_name"
    - "pepl_gender"
    - "pepl_last_name"
    - "pepl_second_name"
    properties:
      pepl_second_name:
        type: "string"
        example: "Иванов"
      pepl_first_name:
        type: "string"
        example: "Иван"
      pepl_last_name:
        type: "string"
        example: "Иванович"
      pepl_gender:
        type: "boolean"
        example: true
      pepl_birthday:
        type: "string"
        format: "date"
        example: "1977-01-19"
      pepl_phone:
        type: "string"
        example: "9568734554"
      pepl_email:
        type: "string"
        example: "admin@mail.ru"
    description: "Данные пользователя"
    example:
      pepl_first_name: "Иван"
      pepl_gender: true
      pepl_second_name: "Иванов"
      pepl_last_name: "Иванович"
      pepl_email: "admin@mail.ru"
      pepl_phone: "9568734554"
      pepl_birthday: "1977-01-19"
...
  Parent:
    type: "object"
    required:
    - "prnt_city"
    - "prnt_home"
    - "prnt_street"
    properties:
      prnt_city:
        type: "string"
        example: "Липецк"
      prnt_street:
        type: "string"
        example: "Ангарская"
      prnt_home:
        type: "string"
        example: "17Б"
      prnt_flat:
        type: "string"
        example: "45"
    description: "Данные родителя"
    example:
      prnt_flat: "45"
      prnt_home: "17Б"
      prnt_city: "Липецк"
      prnt_street: "Ангарская"
...
  Post:
    type: "object"
    required:
    - "pst_id"
    - "pst_name"
    properties:
      pst_id:
        type: "integer"
        format: "int32"
        example: 1
      pst_name:
        type: "string"
        example: "Педагог-психолог"
      pst_description:
        type: "string"
    example:
      pst_id: 1
      pst_name: "Педагог-психолог"
      pst_description: "pst_description"
...
  body:
    required:
    - "file_id"
    properties:
      file_id:
        type: "integer"
        format: "int32"
        example: 3
  inline_response_200_payload:
    required:
    - "file"
    - "file_dt"
    - "file_id"
    - "file_name"
    properties:
      file_id:
        type: "integer"
        format: "int32"
        example: 3
      file_name:
        type: "string"
        example: "Тест по профорейнтации"
      file_path:
        type: "string"
        example: "/folder1"
      file:
        type: "string"
        format: "base64"
      file_dt:
        type: "string"
        format: "date-time"
        example: "2019-03-04T09:35:00.000Z"
...
  body_1:
    type: "object"
    properties:
      std_id:
        type: "integer"
        format: "int32"
        example: 3
      tr_id:
        type: "integer"
        format: "int32"
        example: 1
      tst_id:
        type: "integer"
        format: "int32"
        example: 1
  inline_response_200_1_payload:
    required:
    - "file_dt"
    - "file_id"
    - "file_name"
    properties:
      file_id:
        type: "integer"
        format: "int32"
        example: 3
      file_name:
        type: "string"
        example: "Тест по профорейнтации"
      file_path:
        type: "string"
        example: "/folder1"
      file_dt:
        type: "string"
        format: "date-time"
        example: "2019-03-04T09:35:00.000Z"
    example:
      file_path: "/folder1"
      file_name: "Тест по профорейнтации"
      file_id: 3
      file_dt: "2019-03-04T09:35:00.000Z"
...
  addFiles_file_arr:
    required:
    - "file"
    - "file_name"
    properties:
      file_name:
        type: "string"
        example: "Тест по профорейнтации"
      file_path:
        type: "string"
        example: "/folder1"
      file:
        type: "string"
        format: "base64"
  body_2:
    type: "object"
    required:
    - "file_arr"
    properties:
      std_id:
        type: "integer"
        format: "int32"
        example: 3
      tr_id:
        type: "integer"
        format: "int32"
        example: 1
      tst_id:
        type: "integer"
        format: "int32"
        example: 1
      file_arr:
        type: "array"
        items:
          $ref: "#/definitions/addFiles_file_arr"
  inline_response_200_2_payload:
    required:
    - "file_id"
    - "file_name"
    properties:
      file_id:
        type: "integer"
        format: "int32"
        example: 3
      file_name:
        type: "string"
        example: "Тест по профорейнтации"
    example:
      file_name: "Тест по профорейнтации"
      file_id: 3
...
  body_3:
    type: "object"
    required:
    - "file_id"
    properties:
      file_id:
        type: "integer"
        format: "int32"
        example: 3
      file_name:
        type: "string"
        example: "Тест по профорейнтации"
      file_path:
        type: "string"
        example: "/folder1"
...
  inline_response_200_4_payload:
    required:
    - "file_del_status"
    - "file_id"
    properties:
      file_id:
        type: "integer"
        format: "int32"
        example: 3
      file_del_status:
        type: "boolean"
        example: true
    example:
      file_del_status: true
      file_id: 3
...
  signup_emp_data_pst_arr:
    required:
    - "pst_id"
    properties:
      pst_id:
        type: "integer"
        format: "int32"
        example: 1
...
  inline_response_200_18_payload:
    required:
    - "file"
    properties:
      file:
        type: "string"
        format: "base64"
...
  body_23:
    type: "object"
    required:
    - "prnt_confirm"
    - "prnt_id"
    properties:
      prnt_id:
        type: "integer"
        format: "int32"
        example: 2
      prnt_confirm:
        type: "integer"
        format: "int32"
        example: 1
        description: "Value:\n * 0 - на рассмотрении\n * 1 - подтверждено\n * 2 -\
          \ отклонено\n"
  inline_response_200_19_cr_array:
    required:
    - "cr_class"
    - "cr_first_child"
    - "cr_id"
    - "cr_last_child"
    - "cr_second_child"
    properties:
      cr_id:
        type: "integer"
        format: "int32"
        example: 1
      cr_second_child:
        type: "string"
        example: "Тарасов"
      cr_first_child:
        type: "string"
        example: "Семен"
      cr_last_child:
        type: "string"
        example: "Львович"
      cr_second_teacher:
        type: "string"
        example: "Иванов"
      cr_first_teacher:
        type: "string"
        example: "Иван"
      cr_last_teacher:
        type: "string"
        example: "Иванович"
      cr_class:
        type: "string"
        example: "7Б"
    example:
      cr_first_child: "Семен"
      cr_last_teacher: "Иванович"
      cr_id: 1
      cr_class: "7Б"
      cr_first_teacher: "Иван"
      cr_second_child: "Тарасов"
      cr_last_child: "Львович"
      cr_second_teacher: "Иванов"
  inline_response_200_19_payload:
    required:
    - "cr_array"
    - "pepl_data"
    - "prnt_data"
    - "prnt_id"
    properties:
      prnt_id:
        type: "integer"
        format: "int32"
        example: 2
      cr_array:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_19_cr_array"
      pepl_data:
        $ref: "#/definitions/People"
      prnt_data:
        $ref: "#/definitions/Parent"
    example:
      prnt_id: 2
      cr_array:
      - cr_first_child: "Семен"
        cr_last_teacher: "Иванович"
        cr_id: 1
        cr_class: "7Б"
        cr_first_teacher: "Иван"
        cr_second_child: "Тарасов"
        cr_last_child: "Львович"
        cr_second_teacher: "Иванов"
      - cr_first_child: "Семен"
        cr_last_teacher: "Иванович"
        cr_id: 1
        cr_class: "7Б"
        cr_first_teacher: "Иван"
        cr_second_child: "Тарасов"
        cr_last_child: "Львович"
        cr_second_teacher: "Иванов"
      pepl_data:
        pepl_first_name: "Иван"
        pepl_gender: true
        pepl_second_name: "Иванов"
        pepl_last_name: "Иванович"
        pepl_email: "admin@mail.ru"
        pepl_phone: "9568734554"
        pepl_birthday: "1977-01-19"
      prnt_data:
        prnt_flat: "45"
        prnt_home: "17Б"
        prnt_city: "Липецк"
        prnt_street: "Ангарская"
...
  body_24:
    required:
    - "prnt_id"
    properties:
      prnt_id:
        type: "integer"
        format: "int32"
        example: 1
  inline_response_200_20_payload:
    required:
    - "cr_child_fullname"
    - "cr_class"
    - "cr_id"
    - "cr_teacher_fullname"
    properties:
      cr_id:
        type: "integer"
        format: "int32"
        example: 1
      cr_child_fullname:
        type: "string"
        example: "OK"
      cr_teacher_fullname:
        type: "string"
        example: "OK"
      cr_class:
        type: "string"
        example: "OK"
    example:
      cr_id: 1
      cr_teacher_fullname: "OK"
      cr_class: "OK"
      cr_child_fullname: "OK"
...
  body_33:
    type: "object"
    required:
    - "emp_id"
    properties:
      emp_id:
        type: "integer"
        format: "int32"
        example: 1
...
  body_34:
    type: "object"
    required:
    - "emp_id"
    - "pst_array"
    properties:
      emp_id:
        type: "integer"
        format: "int32"
        example: 1
      pst_array:
        type: "array"
        items:
          $ref: "#/definitions/signup_emp_data_pst_arr"
  body_35:
    type: "object"
    required:
    - "emp_id"
    - "pst_array"
    properties:
      emp_id:
        type: "integer"
        format: "int32"
        example: 1
      pst_array:
        type: "array"
        items:
          $ref: "#/definitions/signup_emp_data_pst_arr"
  body_36:
    required:
    - "pst_name"
    properties:
      pst_name:
        type: "string"
        example: "Педагог-психолог"
      pst_description:
        type: "string"